Working on the ProductsDAO Class
Started work on `ProductsDAO` class and have created a `Task<List<Products>, Void>` for getting all of the records from the table. This will not really need to be done in a task to start, but could quickly be useful for providing feedback to the user that something is happening behind the scenes. I will want to use `Task` subclasses for all getting/saving and searching on the database just to provide feedback to the user. Set up with the SAF, might just as well use it.

-SC