helpLabel Working
I have gotten the `helpLabel` on the `ProductEntry` window working, showing the help text for the fields from the resources file. Had an initial issue of the the label resizing to hold the text, but got that fixed. I had autoresize turned off for both horizontal and vertical, but it still resized horizontally. I set the `preferredSize`, `minimumSize`, and `maximumSize` properties to the same values: width = 369, height = 117. Now it does not autoresize and breaks the lines of the text as it should.

I also investigated how to display images from the database in the `imageLabel`. I will need to use a `BufferedImage` object to do so.

-SC